fix: address cross-owner hash collision in generation pipeline
The async generation pipeline was calling randomUUID() blindly when
picking a row id, missing the cross-owner hash-collision check that
PR #199 added to the CRUD path (pickCollisionFreeId via buildUniqueSlug).

With ~67M values in the 5-char base36 space, the per-owner unique
constraint catches owner-local conflicts but two owners on the same
base slug can mint indistinguishable <base>.<hash> URLs. The resolver
returns 404 on multiple-match, leaving generation-created rows
unreachable via their public URL.

Changes:
- Extract pickCollisionFreeId (formerly inline in create.ts) into
  src/api/v2/agent-templates/lib/pick-collision-free-id.ts so both
  the CRUD and generation paths use the same policy.
- Add isSlugExhaustionError helper so retry loops can recognise
  buildUniqueSlug's 8-attempt-budget-exhausted error and advance to
  the next -N suffix instead of bubbling up.
- generation-executor.persistTemplate now calls pickCollisionFreeId
  per candidate base. Exhaustion (extremely unlikely) advances to
  the next auto-suffix.
