+++ src/api/v2/agent-templates/lib/pick-collision-free-id.ts
@@ -1,0 +1,56 @@
+/**
+ * Pick a row id whose `slugHash(id)` doesn't collide with any existing
+ * AgentTemplate row sharing the given `baseSlug` — ACROSS owners.
+ *
+ * Context: the per-owner DB unique constraint catches owner-local slug
+ * conflicts, but the public hashed URL `<base>.<hash5>` is cross-owner.
+ * Two owners on the same base slug can mint indistinguishable URLs
+ * unless we pre-pick a non-colliding id. With ~67M values in the 5-char
+ * base36 space, collisions are negligible at small scale but grow with
+ * the birthday bound (~1% at 1.1k rows sharing a base, ~50% at 9k).
+ *
+ * Residual race: between this read and the subsequent INSERT, a concurrent
+ * transaction could insert a row whose hash collides with our pick. For
+ * that to produce a real collision, two writes need to land in the same
+ * sub-ms window, share the same base slug, and have UUIDs that hash to
+ * the same 5 chars. Compound probability is negligible at our scale;
+ * revisit with a SERIALIZABLE transaction or a stored urlHash unique
+ * index if collisions ever surface in monitoring.
+ *
+ * Used by both the CRUD create path (handlers/create.ts) and the async
+ * generation pipeline (services/generation-executor.ts) so both surfaces
+ * mint cross-owner-unique URLs by the same policy.
+ */
+
+import { randomUUID } from "node:crypto";
+import { prisma } from "@/utils/prisma";
+import { buildUniqueSlug, slugHash } from "@/utils/slug-hash";
+
+export async function pickCollisionFreeId(args: {
+  baseSlug: string;
+}): Promise<string> {
+  const { id } = await buildUniqueSlug({
+    baseSlug: args.baseSlug,
+    idFactory: () => randomUUID(),
+    isTaken: async (candidate) => {
+      const dot = candidate.lastIndexOf(".");
+      const base = candidate.slice(0, dot);
+      const hash = candidate.slice(dot + 1);
+      const rows = await prisma.agentTemplate.findMany({
+        where: { slug: base },
+        select: { id: true },
+      });
+      return rows.some((row) => slugHash(row.id) === hash);
+    },
+  });
+  return id;
+}
+
+/** Thrown by `buildUniqueSlug` when its retry budget is exhausted.
+ *  Callers (auto-suffix retry loops) should treat this like a slug
+ *  conflict on the current base and try the next `-N` candidate. */
+export function isSlugExhaustionError(err: unknown): boolean {
+  return (
+    err instanceof Error && err.message.startsWith("slug collision: exhausted")
+  );
+}
